requirements.txt
Django==1.9.5
django-appconf==1.0.2
django-compressor==2.0
django-crispy-forms==1.6.0
jsonfield==1.0.3
mysqlclient==1.3.7
rcssmin==1.0.6
rjsmin==1.0.12
